Enhance clinic system: billing, visits, UI, and refactoring
- Added patient check-in dialog with payment method selection
- Integrated billing logic with tax calculation and payment recording
- Refactored visit and vitals models to track creator and use byte status
- Redesigned doctor station UI: waiting list, vitals, validation, new actions
- Added visit services management (retrieve, bulk save)
- Introduced system settings data access for tax rates
- Unified nurse and doctor station data refresh logic
- Improved error handling, logging, and code maintainability throughout
- Updated SQL/data access for new fields and transactional operations

+++ Clinic/Medical Services/Appointment/frmAddUpdateAppointment.cs
@@ -142,7 +142,7 @@ namespace Clinic.Medical_Services.Appointment
             _Appointment.AppointmentFees = Convert.ToDecimal(lblTotalAppointmentFees.Text);
             _Appointment.AppointmentDate = dtpAppointmentDate.Value;
             if (_Mode == enMode.AddNew)
-                _Appointment.AppointmentStatus = (short)clsAppointment.enAppointmentStatus.Scheduled;
+                _Appointment.AppointmentStatus = (byte)clsAppointment.enAppointmentStatus.Scheduled;
         }
 
         private void _TotalAppointment()
